order post 기능 구현.

+++ src/main/java/com/koomineat/koomineat/domain/order/entity/OrderItem.java
@@ -18,14 +18,13 @@ public class OrderItem {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long price;
+    private int price;
 
     private int quantity;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id")
     private Order order;
-
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="menu_item_id")
